+++ include/hc_prof_runtime.h
@@ -1,5 +1,15 @@
 #pragma once
 
+namespace hc {
+// HCC Op IDs enumeration
+enum HSAOpId {
+    HSA_OP_ID_DISPATCH = 0,
+    HSA_OP_ID_COPY = 1,
+    HSA_OP_ID_BARRIER = 2,
+    HSA_OP_ID_NUMBER = 3
+};
+}; // namespace hc
+
 namespace Kalmar {
 namespace CLAMP {
 // Activity profiling primitives
